refactor: merge HtmlTagsPlugin into the basic one
