+++ e2e/cases/html-tags/index.test.ts
@@ -1,4 +1,3 @@
-import path from 'node:path';
 import { expect, test } from '@playwright/test';
 import { build } from '@e2e/helper';
